Created a DownscalingModel, reset the Sigmoid_ext_var option, fix error
Created a DownscalingModel to count the number of Overflow Errors happening in the np.exp function. This error is now partially avoided by switching to float128 for the np.exp() call, then switching back to float64. This fix only works for the Sigmoid option but not for the Sigmoid_ext_var option, which generates much smaller/bigger numbers due to bad fit. The  Sigmoid_ext_var option was reset in a clean way.