Cleaning of gam_on_discharge()
The period of interest is moved as argument of the function. Some periods that bug can now be excluded (defined as argument). R is now in verbose mode. The outliers of (a, b, c, M) parameter sets now have to be removed before the function is called. Erase the conditions based on a anc c.